Extracting common question types in a separate file Adding support for file types - document files and unknown document types for submission details Fixing lint and build errors Extracting pdf related icon in a separate file
