+++ apps/endatix-hub/features/pdf-export/components/icons/index.ts
@@ -1,3 +1,4 @@
 export { default as VideoFileIcon } from './video-file-icon';
 export { default as DocumentFileIcon } from './document-file-icon';
 export { default as UnknownFileIcon } from './unknown-file-icon';
+export { default as MessageSquareTextIcon } from './message-square-text-icon';
