+++ apps/endatix-hub/features/submissions/pdf/pdf-answer-viewer.tsx
@@ -1,26 +1,14 @@
 import React from 'react';
-import { Text, View, StyleSheet, Path, Svg } from '@react-pdf/renderer';
+import { Text, View, StyleSheet } from '@react-pdf/renderer';
 import { Question, QuestionFileModel } from 'survey-core';
 import PdfFileAnswer from './pdf-file-answer';
+import { QuestionType } from '@/lib/questions';
+import { MessageSquareTextIcon } from '@/features/pdf-export/components/icons';
 
 export interface ViewAnswerProps {
   forQuestion: Question;
   panelTitle: string;
   pageBreak: boolean;
-}
-
-export enum QuestionType {
-  Text = 'text',
-  Boolean = 'boolean',
-  Rating = 'rating',
-  Radiogroup = 'radiogroup',
-  Dropdown = 'dropdown',
-  Ranking = 'ranking',
-  Matrix = 'matrix',
-  Comment = 'comment',
-  File = 'file',
-  Video = 'video',
-  Unsupported = 'unsupported',
 }
 
 const PdfAnswerViewer = ({
@@ -113,32 +101,6 @@ const PdfAnswerViewer = ({
     </View>
   );
 
-  const MessageSquareTextIcon = () => (
-    <Svg style={styles.icon}>
-      <Path
-        d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"
-        stroke="currentColor"
-        strokeWidth={2}
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
-      <Path
-        d="M13 8H7"
-        stroke="currentColor"
-        strokeWidth={2}
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
-      <Path
-        d="M17 12H7"
-        stroke="currentColor"
-        strokeWidth={2}
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      />
-    </Svg>
-  );
-
   const renderFileAnswer = () => (
     <View style={styles.fileAnswerContainer} break={pageBreak} wrap={false}>
       <Text style={styles.questionLabel}>
@@ -197,7 +159,7 @@ const PdfAnswerViewer = ({
   }
 };
 
-const styles = StyleSheet.create({
+export const styles = StyleSheet.create({
   fileAnswerContainer: {
     marginBottom: 8,
     padding: 8,
