+++ apps/endatix-hub/features/submissions/pdf/pdf-file-answer.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { View, Text, StyleSheet } from "@react-pdf/renderer";
-import { PdfFileViewer, File } from "./pdf-file-viewer";
+import { PdfFileViewer } from "./pdf-file-viewer";
 import { QuestionFileModel } from "survey-core";
-
+import { File } from "@/lib/questions/file/file-type";
 interface FileAnswerProps {
   question: QuestionFileModel;
 }
