+++ apps/endatix-hub/lib/questions/index.ts
@@ -1,5 +1,6 @@
 // export core questions infrastructure
 export * from './infrastructure/specialized-survey-question';
+export * from './questions-registry';
 
 // export specific questions
 export * from './video/specialized-video.question';
