+++ apps/endatix-hub/lib/questions/questions-registry.ts
@@ -1,0 +1,13 @@
+export enum QuestionType {
+    Text = 'text',
+    Boolean = 'boolean',
+    Rating = 'rating',
+    Radiogroup = 'radiogroup',
+    Dropdown = 'dropdown',
+    Ranking = 'ranking',
+    Matrix = 'matrix',
+    Comment = 'comment',
+    File = 'file',
+    Video = 'video',
+    Unsupported = 'unsupported',
+  }
